Improved cleanup of JACK MIDI sessions that were appeared dangling after stopping a synth connected to an exclusive MIDI port.

+++ mt32emu_qt/src/mididrv/JACKMidiDriver.cpp
@@ -117,6 +117,8 @@ bool JACKMidiDriver::createPort(bool exclusive) {
 }
 
 void JACKMidiDriver::onJACKMidiPortDeleted(MidiSession *midiSession) {
-	deletePort(midiSession);
-	emit midiSessionDeleted(midiSession);
+	if (canDeletePort(midiSession)) {
+		deletePort(midiSession);
+		emit midiSessionDeleted(midiSession);
+	}
 }
